# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Conflict": {
        "additionalProperties": false,
        "properties": {
          "conflict_id": {
            "title": "Conflict Id",
            "type": "string"
          },
          "member_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 2,
            "title": "Member Ids",
            "type": "array"
          },
          "selected_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Id"
          },
          "snapshot_revision": {
            "title": "Snapshot Revision",
            "type": "integer"
          }
        },
        "required": [
          "conflict_id",
          "member_ids",
          "snapshot_revision",
          "selected_id"
        ],
        "title": "Conflict",
        "type": "object"
      },
...
      "Disposition": {
        "enum": [
          "scheduled",
          "excused",
          "cancelled",
          "missed"
        ],
        "title": "Disposition",
        "type": "string"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "current_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Revision"
          },
          "field_errors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Occurrence": {
        "additionalProperties": false,
        "properties": {
          "disposition": {
            "$ref": "#/components/schemas/Disposition"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "occurrence_id": {
            "title": "Occurrence Id",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "temporal_phase": {
            "$ref": "#/components/schemas/TemporalPhase"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "occurrence_id",
          "event_id",
          "version",
          "title",
          "location",
          "temporal_phase",
          "disposition",
          "start_at",
          "end_at"
        ],
        "title": "Occurrence",
        "type": "object"
      },
...
      "Reminder": {
        "additionalProperties": false,
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "event_title": {
            "title": "Event Title",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "minutes_until_start": {
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Minutes Until Start",
            "type": "integer"
          },
          "occurrence_id": {
            "title": "Occurrence Id",
            "type": "string"
          },
          "reminder_id": {
            "title": "Reminder Id",
            "type": "string"
          },
          "schedule_revision": {
            "title": "Schedule Revision",
            "type": "integer"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "reminder_id",
          "occurrence_id",
          "version",
          "schedule_revision",
          "acknowledged_at",
          "event_title",
          "location",
          "start_at",
          "minutes_until_start"
        ],
        "title": "Reminder",
        "type": "object"
      },
...
      "StateResponse": {
        "additionalProperties": false,
        "properties": {
          "active_occurrences": {
            "items": {
              "$ref": "#/components/schemas/Occurrence"
            },
            "title": "Active Occurrences",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/Conflict"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "due_reminders": {
            "items": {
              "$ref": "#/components/schemas/Reminder"
            },
            "title": "Due Reminders",
            "type": "array"
          },
          "next_transition_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Transition At"
          },
          "server_now": {
            "format": "date-time",
            "title": "Server Now",
            "type": "string"
          },
          "state_revision": {
            "title": "State Revision",
            "type": "integer"
          }
        },
        "required": [
          "server_now",
          "state_revision",
          "active_occurrences",
          "due_reminders",
          "conflicts",
          "next_transition_at"
        ],
        "title": "StateResponse",
        "type": "object"
      },
      "TemporalPhase": {
        "enum": [
          "upcoming",
          "active",
          "ended"
        ],
        "title": "TemporalPhase",
        "type": "string"
      },
      "WeatherObservation": {
        "additionalProperties": false,
        "properties": {
          "cloud_cover": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Cover"
          },
          "condition": {
            "anyOf": [
              {
                "enum": [
                  "clear",
                  "partly_cloudy",
                  "overcast",
                  "rain",
                  "fog",
                  "snow"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "kind": {
            "enum": [
              "current",
              "forecast"
            ],
            "title": "Kind",
            "type": "string"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "precipitation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation"
          },
          "valid_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          }
        },
        "required": [
          "kind",
          "condition",
          "cloud_cover",
          "precipitation",
          "visibility",
          "observed_at",
          "valid_until"
        ],
        "title": "WeatherObservation",
        "type": "object"
      },
      "WeatherPlaceChoice": {
        "additionalProperties": false,
        "properties": {
          "admin1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin1"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude",
          "timezone",
          "country",
          "admin1"
        ],
        "title": "WeatherPlaceChoice",
        "type": "object"
      },
      "WeatherResponse": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "availability": {
            "enum": [
              "available",
              "stale",
              "unavailable"
            ],
            "title": "Availability",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "fetched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "location_choices": {
            "items": {
              "$ref": "#/components/schemas/WeatherPlaceChoice"
            },
            "title": "Location Choices",
            "type": "array"
          },
          "location_id": {
            "title": "Location Id",
            "type": "string"
          },
          "location_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Label"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/WeatherObservation"
            },
            "title": "Observations",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "required": [
          "location_id",
          "availability",
          "source",
          "fetched_at",
          "observations"
        ],
        "title": "WeatherResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kairos API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/state": {
      "get": {
        "operationId": "state_api_v1_state_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "State"
      }
    },
    "/api/v1/weather": {
      "get": {
        "operationId": "weather_api_v1_weather_get",
        "parameters": [
          {
            "in": "query",
            "name": "location_id",
            "required": true,
            "schema": {
              "title": "Location Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Weather"
      }
    }
  }
}